Rename README.md to readme edits
disaster management can be defined as the organisation and management of resources and resposiblities for dealing with all humanatirian aspects of emergencies in particular preparadness, response and recovery in order to lessen the impact. 
the first people to respond to disaster are those living in the local community.they are the first rescue and relieve oprations. the red cross and red crescent national societies terefore focus on community based disaster preparedness which assist communities to reduce there vulnerability to disasters and strenthen their capacities to resist them.